Updated youtube handler to use "?list=" and "?v=" parameters
Updated parser to allow "youtube.com" and "youtu.be" links to be used

Updated youtube handler to parse out playlist IDs from "list=" using the
SCM.parseYoutubePlaylistId() function and to handle single song requests
if a "v=" is found without a "list=" or "p=" .

Added property 'access' to playlist object.  This properly is "" unless
youtube signifies that it cannot be embeded, then its value is
"restricted".  This currently does not change any functionality of the
code.